% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{download_model}
\alias{download_model}
\title{Download MegaDetector Model}
\usage{
download_model(model = "MDV5A", force = FALSE)
}
\arguments{
\item{model}{Model identifier (default "MDV5A").}

\item{force}{If TRUE, force download even if the file exists.}
}
\value{
Path to the downloaded model file.
}
\description{
Downloads a MegaDetector model to the user's cache directory.
}
